x86/asm: Introduce static_retcall(s)
Provide a framework to overwrite a tail call in a function with return.
For timens case there is no need in reverse patching, so keeping it
simpler and omitted the code to save function call instruction(s).
At this state it's valid only for patching userspace as it has
no protection for parallel execution of patched code.

In comparison to static_branch():
- doesn't need no-op
- no need for architecture-specific part for NOP/JMP instructions
- probably depends much on compiler and flags

Co-developed-by: Andrey Vagin <avagin@openvz.org>
Signed-off-by: Dmitry Safonov <dima@arista.com>