ptrace: Introduce PTRACE_GET_RESTART_TIMEOUT
In CRIU (Checkpoint Restore In Userspace) project it's not possible to
use restart_syscall() to resume a syscall with timeout because
the new [recreated] task_struct doesn't have restart_block set.
Instead, if the task was interrupted on checkpoint, on restore
the instruction pointer is rewind back on syscall instruction
and the register values are set back where they were.

That results in awful user-visible issue: syscalls with a timeout are
restarted with the original timeout. And even worse: if an application
or container is C/R'ed with somewhat "high" frequency - the started
syscall may never return [1].

Provide a new ptrace() API to get timeout from the restart_block on the
task_struct being checkpointed. It seems that not only CRIU will benifit
from PTRACE_GET_RESTART_TIMEOUT, but strace could also add the timeout
to the information it prints when the tracee is signaled.

Unfortunately, the new API in RFC version is incomplete: if the syscall
was interrupted not by CRIU but some time before a checkpoint
(i.e., the application was running in a signal handler), it wouldn't be
possible to restore restart_block members.

[1] https://lists.openvz.org/pipermail/criu/2019-June/044086.html
Not-Yet-Signed-off-by: Dmitry Safonov <dima@arista.com>