[draft] net/tcp: Prepare tcp_md5sig_pool for TCP-AO
TCP-AO similarly to TCP-MD5 needs to allocate tfms on a slow-path, which
is setsockopt() and use crypto ahash requests on fast paths, which are
RX/TX softirqs. It as well needs a temporary/scratch buffer for
preparing the hashing request.

Extend tcp_md5sig_pool to support other hashing algorithms than MD5.
Move it in a separate file.

This patch was previously submitted as more generic crypto_pool [1],
but Herbert nacked making it generic crypto API. His view is that crypto
requests should be atomically allocated on fast-paths. So, in this
version I don't move this pool anywhere outside TCP, only extending it
for TCP-AO use-case. It can be converted once there will be per-request
hashing crypto keys.

NOTE:
This is a draft version of the patch: in discussion with Herbert [2]
he asks to wait for his per-request setkey changes, which he expects
to be ready during v6.3 development (to be in v6.4 kernel).
When they will be in linux-next, I'm going to rework this patch to use
one tfm for each hashing algo with crypto requests dynamically allocated
in RX/TX using kmalloc().
Sending this patch to make buildbots happy as well as for anyone who
wants to try/test TCP-AO.

[1]: https://lore.kernel.org/all/20230118214111.394416-1-dima@arista.com/T/#u
[2]: https://lore.kernel.org/all/c4224fa7-c206-17d3-641b-6f3f53dd813d@arista.com/T/#u
Signed-off-by: Dmitry Safonov <dima@arista.com>